pyproject.toml
[project]
name = "medical-ai-midterm-project"
version = "0.1.0"
description = "Multimodal fusion model for skin lesion classification"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "timm>=1.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    "Pillow>=10.0.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "streamlit>=1.40.0",
    "grad-cam>=1.5.0",
    "tqdm>=4.66.0",
    "albumentations>=1.4.0",
    "kagglehub>=1.0.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu124", marker = "sys_platform == 'win32' or sys_platform == 'linux'" },
]
torchvision = [
    { index = "pytorch-cu124", marker = "sys_platform == 'win32' or sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
